pasando el proyectoa asincrono

+++ Registro_estacionamiento/Registro_estacionamiento/Models/RegistrosModel.cs
@@ -6,8 +6,8 @@
         public int VehiculoId { get; set; }
         public DateTime FechaHoraEntrada { get; set; }
         public DateTime? FechaHoraSalida { get; set; }
-        public decimal CostoPorHora { get; set; }
-        public decimal? CostoTotal { get; set; }
+        public double CostoPorHora { get; set; }
+        public double? CostoTotal { get; set; }
 
         public VehiculosModel VehiculosModel { get; set; }
     }
